created a file that has the api code
for searching in maps later you'll have to enter my address to make the results as live as possible even though its mimicked through already created data. btw you guys should watch on netflix the billion dollar code it really got me emotional. so these guys were building google maps before it existed. anyways the point is the mimicked proving that the code worked when it really didn't but as long as the client believes it thats what matters.